Add delete note endpoint
Notice that the promise isn’t `then`, but `remove`, and that I had to add “DELETE” to the list of allowed HTTP verbs.

If we didn’t want to mess with this middleware ourselves, we could use the npm module https://github.com/expressjs/cors to handle it all for us.